integer time

+++ import.R
@@ -1,7 +1,7 @@
 files <- list.files()
 
 readfile <- function(name) {
-  data <- read.table(name, na.strings="None", colClasses=c("numeric", "numeric"), col.names=c("time", name))
+  data <- read.table(name, na.strings="None", colClasses=c("integer", "numeric"), col.names=c("time", name))
   data <- data[!is.na(data[[2]]),]
 }
